requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
requests==2.31.0
python-telegram-bot==20.6
openpyxl==3.1.2
pandas==2.1.4